requirements.txt
openai==1.14.3
Pillow==9.5.0
Pillow==10.2.0
pydub==0.25.1
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.32.2
